Fix voices list parsing
Apparently csv, the bundled Ruby gem introduced some change in a 2.6.x
version that breaks the parsing of the espeak --voices output. To be
fair, csv parser was possible not the best choice in the first place.

Also remove usage of the fixture from the test to catch regression in
case of espeak changing the format of the output.